feat: add HemeraTermFx (formerly HemeraRenderer/HemeraTerm) to print Aether's final, merged frame to the terminal. #4, #5, #7
